requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
requests
beautifulsoup4
feedparser
python-dotenv
fastapi
uvicorn
rich
pydantic>=2.0
